+++ .codeclimate.yml
@@ -66,5 +66,5 @@ exclude_patterns:
 # These files were downloaded from https://designsystem.digital.gov/
 - app/assets/stylesheets/uswds.css
 - app/assets/stylesheets/uswds_override.css
-# This file was downloaded from https://github.com/trussworks/react-uswds/blob/main/src/components/Pagination/Pagination.tsx
+# Overriding USWDS components
 - app/javascript/components/UswdsOverrides/
